[bugfix] use fp32 for rollout_log_probs (#245)

+++ slime/backends/utils/data.py
@@ -145,7 +145,7 @@ def process_rollout_data(args, rollout_data_ref, dp_rank, dp_size):
             torch.tensor(
                 slice_log_prob_with_cp(log_prob, total_length, response_length),
                 device=torch.cuda.current_device(),
-                dtype=torch.bfloat16,  # TODO: hardcode to bf16 at the moment
+                dtype=torch.float32,
             )
             for log_prob, total_length, response_length in zip(
                 rollout_data["rollout_log_probs"], rollout_data["total_lengths"], rollout_data["response_lengths"]
